Transfer functions in utils folders to separate folders
