+++ src/app/bid/types.go
@@ -1,0 +1,6 @@
+package bid
+
+type BidAPI struct {
+	ItemID   uint32 `json:"itemId"`
+	BidPrice string `json:"bidPrice"`
+}
