--- src/app/item/types.go
@@ -1,1 +1,0 @@
-package item
