+++ src/app/user/types.go
@@ -1,0 +1,9 @@
+package user
+
+type LoginCredentialsAPI struct {
+	Email     string
+	Password  string
+	IsSignUp  bool
+	FirstName string
+	LastName  string
+}
